+++ sdk/resourcemanager/containerservice/armcontainerservice/trustedaccessroles_client_example_test.go
@@ -11,7 +11,7 @@ import (
 	"log"
 )
 
-// Generated from example definition: 2026-01-01/TrustedAccessRoles_List.json
+// Generated from example definition: 2026-01-02-preview/TrustedAccessRoles_List.json
 func ExampleTrustedAccessRolesClient_NewListPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {