+++ sdk/resourcemanager/search/armsearch/fake/operations_server.go
@@ -7,13 +7,12 @@ package fake
 import (
 	"errors"
 	"fmt"
-	"net/http"
-
 	azfake "github.com/Azure/azure-sdk-for-go/sdk/azcore/fake"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/fake/server"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/search/armsearch/v2"
+	"net/http"
 )
 
 // OperationsServer is a fake server for instances of the armsearch.OperationsClient type.